+++ src/App.css
@@ -1,3 +1,4 @@
 .App {
   text-align: center;
+  font-family: 'Lato', sans-serif;
 }
