+++ src/App.js
@@ -2,8 +2,10 @@ import './App.css';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Home from './components/Home';
 import Details from './components/Details';
+import Header from './components/Header';
 import store from './redux/cofigureStore';
 import { getAnime } from './redux/api/api';
+import 'bootstrap/dist/css/bootstrap.min.css';
 
 store.dispatch(getAnime());
 
@@ -11,6 +13,7 @@ function App() {
   return (
     <div className="App">
       <Router>
+        <Header />
         <Switch>
           <Route path="/" exact component={Home} />
           <Route path="/details/:id" exact component={Details} />
