+++ src/components/style/Details.css
@@ -1,0 +1,7 @@
+.animeInfo {
+  padding: 0;
+  margin: 0;
+  background-color: #4369b2;
+  width: 100%;
+  height: 100%;
+}
