--- src/components/style/Home.css
@@ -1,48 +1,0 @@
-.dropbtn {
-  background-color: #e45454;
-  width: 100%;
-  height: 100%;
-  color: white;
-  padding: 16px;
-  font-size: 16px;
-  border: none;
-  cursor: pointer;
-}
-
-.dropdown {
-  position: relative;
-  display: inline-block;
-  width: 25vw;
-}
-
-.dropdown-content {
-  width: 100%;
-  height: 100%;
-  display: none;
-  position: absolute;
-  background-color: #f9f9f9;
-  min-width: 160px;
-  z-index: 1;
-}
-
-.dropdown-content button {
-  color: black;
-  border: none;
-  text-decoration: none;
-  display: block;
-  width: 100%;
-  height: 100%;
-  cursor: pointer;
-}
-
-.dropdown-content button:hover {
-  background-color: #f1f1f1;
-}
-
-.dropdown:hover .dropdown-content {
-  display: block;
-}
-
-.dropdown:hover .dropbtn {
-  background-color: #3e8e41;
-}
